for validations

+++ CMS Api-Raag/CMS Api-Raag/ViewModel/PrescriptionmedicinesViewModel.cs
@@ -12,6 +12,6 @@ namespace CMS_Api_Raag.ViewModel
         public string Medicinename { get; set; }
         public string Dosage { get; set; }
         public int Quantity { get; set; }
-        public int PriceperUnit { get; set; }
+        public decimal PriceperUnit { get; set; }
     }
 }
